compiler: more refactoring
