+++ tests/printer_test.py
@@ -1,12 +1,14 @@
 from __future__ import annotations
 
 from io import StringIO
+from typing import List
 
 from xdsl.dialects.builtin import Builtin, IntAttr, ModuleOp, IntegerType, UnitAttr
 from xdsl.dialects.arith import Arith, Addi, Constant
-from xdsl.diagnostic import Diagnostic
-from xdsl.ir import Attribute, MLContext, ParametrizedAttribute
-from xdsl.irdl import ParameterDef, irdl_attr_definition, irdl_op_definition, Operation, OperandDef, ResultDef, OptAttributeDef
+
+from xdsl.ir import Attribute, MLContext, ParametrizedAttribute, Diagnostic
+from xdsl.irdl import (ParameterDef, irdl_attr_definition, irdl_op_definition, Operation,
+                       OperandDef, ResultDef, OptAttributeDef)
 from xdsl.printer import Printer
 from xdsl.parser import Parser
 
@@ -569,7 +571,7 @@ builtin.module() {
     assert file.getvalue().strip() == expected.strip()
 
 
-def test_parse_generic_format_attr():
+def test_parse_generic_format_attr_II():
     """
     Test that we can parse attributes using generic formats.
     """
