+++ xdsl/printer.py
@@ -4,12 +4,11 @@ from dataclasses import dataclass, field
 from enum import Enum
 from frozenlist import FrozenList
 
-from xdsl.diagnostic import Diagnostic
 from typing import Iterable, TypeVar, Any, Dict, Optional, List, cast
 
 from xdsl.dialects.memref import MemRefType
 from xdsl.ir import (BlockArgument, MLIRType, SSAValue, Block, Callable,
-                     Attribute, Region, Operation, ParametrizedAttribute)
+                     Attribute, Region, Operation, Diagnostic)
 from xdsl.dialects.builtin import (
     AnyIntegerAttr, AnyFloatAttr, AnyUnrankedTensorType, AnyVectorType,
     DenseIntOrFPElementsAttr, Float16Type, Float32Type, Float64Type, FloatAttr,
