+++ xdsl/xdsl_opt_main.py
@@ -3,7 +3,7 @@ import sys
 import os
 from io import IOBase, StringIO
 
-from xdsl.ir import MLContext
+from xdsl.ir import MLContext, DiagnosticException
 from xdsl.parser import Parser
 from xdsl.printer import Printer
 from xdsl.dialects.func import Func
@@ -14,7 +14,7 @@ from xdsl.dialects.builtin import ModuleOp, Builtin
 from xdsl.dialects.cmath import CMath
 from xdsl.dialects.cf import Cf
 from xdsl.dialects.memref import MemRef
-from xdsl.diagnostic import DiagnosticException
+from xdsl.exceptions import DiagnosticException
 from xdsl.dialects.llvm import LLVM
 from xdsl.dialects.irdl import IRDL
